tests: Group related functions into a class
Everything related to the O&M interface require the SMSC
OM component to run. By grouping it in a class we can use
a fixture for all the tests without cluttering the method
signature.

Make sure to have "self" as first parameter, thanks to my
debug teddy to find that stupid issue.