manual: Initial version of compiling the manual
Create a new top level introduction. I tried to include the README
as a "developer notes" but leveloffset doesn't work and I can not
re-write the title of the document.

Start describing the GC which is not implemented yet. The highlight
is that there is no way to know which SMS where removed in the end.

Include the PNG for now, at least this reminds me to update it
after changing the dot file.